+++ src/MWVisitor.php
@@ -64,11 +64,10 @@ class MWVisitor extends TaintednessVisitor {
 
 		// Should this be getDefiningFQSEN() instead?
 		$methodName = (string)$method->getFQSEN();
-		$parserFQSEN = MediaWikiHooksHelper::getInstance()->getMwParserClassFQSEN( $this->code_base )->__toString();
 		// $this->debug( __METHOD__, "Checking to see if we should register $methodName" );
 		switch ( $methodName ) {
-			case "$parserFQSEN::setFunctionHook":
-			case "$parserFQSEN::setHook":
+			case "\\MediaWiki\\Parser\\Parser::setFunctionHook":
+			case "\\MediaWiki\\Parser\\Parser::setHook":
 				$type = $this->getHookTypeForRegistrationMethod( $methodName );
 				if ( $type === null ) {
 					break;
@@ -287,11 +286,10 @@ class MWVisitor extends TaintednessVisitor {
 	 * @return string|null The name of the hook that gets registered
 	 */
 	private function getHookTypeForRegistrationMethod( string $method ): ?string {
-		$parserFQSEN = MediaWikiHooksHelper::getInstance()->getMwParserClassFQSEN( $this->code_base )->__toString();
 		switch ( $method ) {
-			case "$parserFQSEN::setFunctionHook":
+			case "\\MediaWiki\\Parser\\Parser::setFunctionHook":
 				return '!ParserFunctionHook';
-			case "$parserFQSEN::setHook":
+			case "\\MediaWiki\\Parser\\Parser::setHook":
 				return '!ParserHook';
 			default:
 				$this->debug( __METHOD__, "$method not a hook registerer" );