+++ tests/integration/stripitem/test.php
@@ -2,13 +2,13 @@
 
 class SomeClass {
 	public function register() {
-		$parser = new Parser;
+		$parser = new MediaWiki\Parser\Parser();
 
 		$parser->setFunctionHook( 'something', [ __CLASS__, 'xyzzy' ] );
 		$parser->setFunctionHook( 'something', [ __CLASS__, 'xyzzy2' ] );
 	}
 
-	public function xyzzy( Parser $parser, $arg1 ) {
+	public function xyzzy( MediaWiki\Parser\Parser $parser, $arg1 ) {
 		$a = $parser->insertStripItem( $arg1 );
 		return $a;
 	}